Automatically handles textures Transparency and Overlay
- Determines at runtime depending on texture it's transparency (Opaque, Cutout, Transparent)
- Determine if Overlay based on texture Model

+++ src/client/src/renderer/world_renderer/block_registry/BlockModel.hpp
@@ -48,8 +48,8 @@ namespace onion::voxel
 
 		struct Element
 		{
-			std::array<float, 3> From{0, 0, 0};
-			std::array<float, 3> To{0, 0, 0};
+			std::array<uint8_t, 3> From{0, 0, 0};
+			std::array<uint8_t, 3> To{0, 0, 0};
 
 			std::unordered_map<std::string, Face> Faces;
 			// keys: "north", "south", etc.
